fixed diary entries for psy
