+++ frontend/src/components/psy/DiaryNotes/DiaryNotes.tsx
@@ -20,6 +20,7 @@ export default function DiaryNotes({ clientId }) {
     setLimit((prev) => prev + 20);
   }, []);
 
+
   useObserve(observeElement, isTotal, handleTakeUpdate);
 
   React.useEffect(() => {
