+++ frontend/src/components/psy/DiaryPageContent/DiaryBlockAnalysis/DiaryBlockAnalysis.tsx
@@ -2,8 +2,11 @@
 
 import React from "react";
 import "../DiaryPageContent.css";
+import { parseEventDetailsText } from "../../../../utils/helperFunction/parseEventDetailsText";
 
 export default function DiaryBlockAnalysis({ diary }) {
+
+
   return (
     <div className="diary__block-event">
       <div className="diary__block-title">Thoughts Analysis</div>
@@ -11,7 +14,7 @@ export default function DiaryBlockAnalysis({ diary }) {
         Reflect on your thoughts related to the situation. What were you
         thinking?
       </div>
-      <div className="diary__block-description">{diary.thoughts_analysis}</div>
+      <div className="diary__block-description">{parseEventDetailsText(diary.thoughts_analysis)}</div>
     </div>
   );
 }
