+++ frontend/src/components/psy/DiaryPageContent/DiaryEventDetails/DiaryEventDetails.tsx
@@ -1,6 +1,7 @@
 //@ts-nocheck
 import React from "react";
 import "../DiaryPageContent.css";
+import { parseEventDetailsText } from "../../../../utils/helperFunction/parseEventDetailsText";
 
 export default function DiaryEventDetails({ diary }) {
   return (
@@ -9,7 +10,7 @@ export default function DiaryEventDetails({ diary }) {
       <div className="diary__block-question">
         Describe the event or situation that evoked emotions. What happened?
       </div>
-      <div className="diary__block-text">{diary.event_details}</div>
+      <div className="diary__block-text"> {parseEventDetailsText(diary.event_details)}</div>
     </div>
   );
 }
